player movement file cleanup
moved things to different files == cleaner
